checkpoint: extracts "Version1/definition.ts"
